+++ Vector2.h
@@ -3,11 +3,11 @@
 
 struct Vector2
 {
-    float x;
-    float y;
+    int x;
+    int y;
 
     Vector2();
-    Vector2(float x, float y);
+    Vector2(int x, int y);
     ~Vector2() = default;
 };
